pyproject.toml
[project]
name = "esm2-contact-prediction"
version = "0.1.0"
description = "ESM2-based protein contact prediction with structural data integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.80",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "ipywidgets>=8.0.0",
    "jupyter>=1.1.1",
    "h5py>=3.15.1",
    "torch>=2.9.0",
    "requests>=2.31.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.7.2",
    "psutil>=7.1.1",
    "ipykernel>=7.0.1",
    "fair-esm>=2.0.0",
    "mlflow>=2.15.0",
    "optuna>=3.6.0",
    "hyperopt>=0.2.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/esm2_contact"]

[project.scripts]
process-dataset = "scripts.process_dataset:main"
test-pipeline = "scripts.test_pipeline:test_processing_pipeline"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
